refactor : Split Zustand store structure and optimize search feature
