+++ src/app/travel/plans/page.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import SearchBar from "@/components/SearchBar";
 import TripCard from "@/components/TripCard";
 import Category from "@/components/Category";
-import { apiClient } from "@/service/interceptor";
 
 async function getPlans() {
   // try {
