+++ src/store/store.tsx
@@ -3,22 +3,6 @@ import { devtools, persist } from "zustand/middleware";
 import type {} from "@redux-devtools/extension"; // required for devtools typing
 import LocalStorage from "@/service/localstorage";
 
-type List = {
-  id: number;
-  createdAt: string;
-  updatedAt: string;
-  title: string;
-  content: string;
-  travelArea: string;
-  travelerCount: number;
-  budget: number;
-  thumbnail: string;
-  startDate: string;
-  endDate: string;
-  visibilityStatus: string;
-  travelPostImage?: [];
-};
-
 interface BearState {
   accessToken: string;
   setAccessToken: (token: string) => void;
@@ -26,12 +10,6 @@ interface BearState {
   setPlanId: (id: number | null) => void;
   isLogin: string;
   setIsLogin: (loginState: string) => void;
-  plans: List[] | [];
-  setPlans: (plans: any) => void;
-  selectedCategory: string;
-  setSelectedCategory: (category: string) => void;
-  searchKeyword: string;
-  setSearchKeyword: (keyword: string) => void;
 }
 
 const access = LocalStorage.getItem("accessToken");
@@ -46,13 +24,6 @@ const useStore = create<BearState>()(
         setPlanId: (id: number | null) => set(() => ({ planId: id })),
         isLogin: "false",
         setIsLogin: (loginState) => set(() => ({ isLogin: loginState })),
-        plans: [],
-        setPlans: (plans) => set(() => ({ plans: plans })),
-        selectedCategory: "",
-        setSelectedCategory: (category) =>
-          set(() => ({ selectedCategory: category })),
-        searchKeyword: "",
-        setSearchKeyword: (keyword) => set(() => ({ searchKeyword: keyword })),
       }),
       {
         name: "bear-storage",
